fix(triton): add autograd wrappers for RMSNorm and geometric_product
Triton kernels broke autograd, causing only embed.weight to receive gradients.
Added torch.autograd.Function wrappers that use Triton for forward and
manual PyTorch backward for gradient correctness.

Also fixed MoE scatter dtype mismatch under fp16 autocast.

Throughput: 5217 tokens/sec (vs 18000 with broken gradients, vs 2591 with
pure PyTorch fallback). Training now produces correct gradients.

+++ src/hagi/model/moe.py
@@ -54,7 +54,7 @@ class MoESwiGLU(nn.Module):
         top_k_probs, top_k_indices = torch.topk(router_probs, self.top_k, dim=-1)
         top_k_probs = top_k_probs / top_k_probs.sum(dim=-1, keepdim=True)
 
-        weight_matrix = torch.zeros(B * T, self.num_experts, device=x.device, dtype=x.dtype)
+        weight_matrix = torch.zeros(B * T, self.num_experts, device=x.device, dtype=top_k_probs.dtype)
         weight_matrix.scatter_(1, top_k_indices, top_k_probs)
 
         output = torch.zeros_like(flat)
